moved auxiliary files to avr_aux
